# openapi-chatgpt-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Breezeway Property API",
    "version": "1.0.0",
    "description": "Access JZ Vacation Rentals property data from Breezeway"
  },
  "servers": [
    {
      "url": "https://breezeway-mcp-server-production.up.railway.app"
    }
  ],
  "paths": {
    "/api/properties/search": {
      "get": {
        "operationId": "searchProperties",
        "summary": "Search properties",
        "description": "Search for properties by name, address, or other attributes",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/properties/internal/{id}": {
      "get": {
        "operationId": "getPropertyByInternalId",
        "summary": "Get property by Streamline ID",
        "description": "Get property details using Streamline internal ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Streamline property ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/properties/{id}": {
      "get": {
        "operationId": "getPropertyById",
        "summary": "Get property by Breezeway ID",
        "description": "Get detailed property information",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Breezeway property ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/properties": {
      "get": {
        "operationId": "listProperties",
        "summary": "List all properties",
        "description": "Get all Breezeway properties",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "operationId": "listTasks",
        "summary": "List tasks",
        "description": "Get tasks with optional filters",
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "description": "Filter by property ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reservations": {
      "get": {
        "operationId": "listReservations",
        "summary": "List reservations",
        "description": "Get reservations with optional filters",
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "description": "Filter by property ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "description": "Filter by start date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "description": "Filter by end date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}